requirements/base.txt
django>=4.2
djangorestframework>=3.14
django-cors-headers>=4.3
drf-spectacular>=0.27
pymongo>=4.6.0
selenium==4.16.0
webdriver-manager==4.0.1
requests>=2.31.0
python-dotenv==1.0.0
pyyaml==6.0.1
openpyxl>=3.1.0
